fixed the exception of geting unknown customer

+++ app/Http/Controllers/customerController.php
@@ -207,9 +207,13 @@ class customerController extends Controller
     {
 
         $v=User::find($id);
-        $v->delete();
-        $v=Customer::find($id);
-        $v->delete();
+        if($v!=null){//checking is user exist for Id
+            $v->delete();
+        }
+        $c=Customer::find($id);
+        if($c!=null){//checking is customer exist for Id
+            $c->delete();
+        }
         Alert::success('Deleted successfully.','Done!');
         return redirect('customers');
     }
